+++ Procfile
@@ -2,7 +2,7 @@ friend_collector: python -m app.friend_collection.batch_per_thread
 bq_grapher: python -m app.friend_graphs.bq_grapher
 bq_list_grapher: python -m app.friend_graphs.bq_list_grapher
 bq_weekly_rt_grapher: python -m app.retweet_graphs.bq_weekly_grapher
-graph_analyzer: python -m app.graph_analyzer
+graph_analyzer: python -m app.friend_graphs.graph_analyzer
 pg_migrate: python -m app.models
 pg_pipeline_user_friends: python -m app.pg_pipeline.user_friends
 pg_pipeline_user_details: python -m app.pg_pipeline.user_details
