+++ app/botcode/classifier.py
@@ -4,7 +4,7 @@ import os
 
 from pandas import DataFrame
 
-from app.graph_analyzer import GraphAnalyzer
+from app.friend_graphs.graph_analyzer import GraphAnalyzer
 from app.decorators.number_decorators import fmt_n, fmt_pct
 from app.botcode.investigation import classify_bot_probabilities
