+++ app/bq_service.py
@@ -5,7 +5,6 @@ from google.cloud import bigquery
 
 from app import APP_ENV, seek_confirmation
 from app.decorators.number_decorators import fmt_n
-from app.decorators.datetime_decorators import dt_to_date
 
 load_dotenv()
 
@@ -491,29 +490,6 @@ class BigQueryService():
         """
         return self.execute_query(sql)
 
-
-class RetweetWeek:
-    def __init__(self, row):
-        """
-        A decorator for the rows returned by the fetch_retweet_weeks() query.
-
-        Param row (google.cloud.bigquery.table.Row)
-        """
-        self.row = row
-
-    @property
-    def week_id(self):
-        return f"{self.row.year}-{str(self.row.week).zfill(2)}" #> "2019-52", "2020-01", etc.
-
-    @property
-    def details(self):
-        details = ""
-        details += f"ID: {self.week_id} | "
-        details += f"FROM: '{dt_to_date(self.row.min_created)}' "
-        details += f"TO: '{dt_to_date(self.row.max_created)}' | "
-        details += f"DAYS: {fmt_n(self.row.day_count)} | "
-        details += f"USERS: {fmt_n(self.row.user_count)} | " + f"RETWEETS: {fmt_n(self.row.retweet_count)}"
-        return details
 
 if __name__ == "__main__":
